Changed the deposits features.
