+++ resources/views/tontine/pages/meeting/deposit/pool/home.blade.php
